Use charlock_holmes
charlock_holmes is a gem that wraps [ICU](http://site.icu-project.org/). It is very good at guessing the encoding of a string through heuristics. If it can detect an encoding, use it – or else take our current nuclear approach.

Also, `encode!` → `encode`: don’t modify the string in place.